Added Completions

+++ open_ai_token/crud.py
@@ -1,21 +1,46 @@
 from sqlalchemy.orm import Session
 
 from open_ai_token import models, schemas
+import uuid
 
+def check_token(db: Session, token: str):
+    que = db.query(models.Token).filter(models.Token.token == token)
+    if que.count() == 0:
+        return False
+    return True
 
 def get_user_by_slack_id(db: Session, slack_id: str):
-    return db.query(models.User).filter(models.User.slack_id == slack_id).first()
+    que = db.query(models.User).filter(models.User.slack_id == slack_id)
+    if que.count() == 0:
+        return None
+    return que.first()
 
 def get_user_by_email(db: Session, email: str):
-    return db.query(models.User).filter(models.User.email == email).first()
+    que = db.query(models.User).filter(models.User.email == email)
+    if que.count() == 0:
+        return None
+    return que.first()
 
 def get_user_by_slack_id_and_email(db: Session, slack_id: str, email: str):
-    return db.query(models.User).filter(models.User.slack_id == slack_id, models.User.email == email).first()
+    que = db.query(models.User).filter(models.User.slack_id == slack_id, models.User.email == email)
+    if que.count() == 0:
+        return None
+    return que.first()
+
 
 def get_users(db: Session, skip: int = 0, limit: int = 100):
-    return db.query(models.User).offset(skip).limit(limit).all()
+    if limit > 100:
+        limit = 100
+    if skip < 0:
+        skip = 0
+    que = db.query(models.User).offset(skip).limit(limit)
+    return que.all()
 
 def create_user(db: Session, user: schemas.UserCreate):
+    # check if user exists
+    db_user = db.query(models.User).filter(models.User.slack_id == user.slack_id).first()
+    if db_user:
+        raise ValueError("User already exists")
     db_user = models.User(slack_id=user.slack_id, name=user.name, email=user.email)
     db.add(db_user)
     db.commit()
@@ -29,20 +54,33 @@ def get_tokens(db: Session, skip: int = 0, limit: int = 100):
     return db.query(models.Token).offset(skip).limit(limit).all()
 
 def create_token(db: Session, token: schemas.TokenCreate):
-    db_token = models.Token(token=token.token, owner_slack_id=token.owner_slack_id)
+    tk = uuid.uuid4()
+    if check_token(db, str(tk)):
+        raise ValueError("Token already exists")
+
+    db_token = models.Token(token=tk, owner_slack_id=token.owner_slack_id)
     db.add(db_token)
     db.commit()
     db.refresh(db_token)
     return db_token
 
 def use_token(db: Session, token: schemas.TokenUse):
+    if not check_token(db, token.token):
+        raise ValueError("Token does not exist")
+
     db_token = db.query(models.Token).filter(models.Token.token == token.token, models.Token.owner_slack_id == token.owner_slack_id).first()
+    if db_token.uses_left == 0:
+        raise ValueError("Token has no uses left")
+
     db_token.uses_left -= 1
     db.commit()
     db.refresh(db_token)
     return db_token
 
 def revoke_token(db: Session, token: schemas.TokenUse):
+    if not check_token(db, token.token):
+        raise ValueError("Token does not exist")
+
     db_token = db.query(models.Token).filter(models.Token.token == token.token, models.Token.owner_slack_id == token.owner_slack_id).first()
     db_token.is_revoked = True
     db.commit()
@@ -50,6 +88,9 @@ def revoke_token(db: Session, token: schemas.TokenUse):
     return db_token
 
 def block_token(db: Session, token: schemas.TokenUse):
+    if not check_token(db, token.token):
+        raise ValueError("Token does not exist")
+
     db_token = db.query(models.Token).filter(models.Token.token == token.token, models.Token.owner_slack_id == token.owner_slack_id).first()
     db_token.is_blocked = True
     db.commit()
@@ -57,6 +98,9 @@ def block_token(db: Session, token: schemas.TokenUse):
     return db_token
 
 def unblock_token(db: Session, token: schemas.TokenUse):
+    if not check_token(db, token.token):
+        raise ValueError("Token does not exist")
+
     db_token = db.query(models.Token).filter(models.Token.token == token.token, models.Token.owner_slack_id == token.owner_slack_id).first()
     db_token.is_blocked = False
     db.commit()
@@ -64,18 +108,27 @@ def unblock_token(db: Session, token: schemas.TokenUse):
     return db_token
 
 def delete_token(db: Session, token: schemas.TokenUse):
+    if not check_token(db, token.token):
+        raise ValueError("Token does not exist")
+
     db_token = db.query(models.Token).filter(models.Token.token == token.token, models.Token.owner_slack_id == token.owner_slack_id).first()
     db.delete(db_token)
     db.commit()
     return db_token
 
 def delete_user(db: Session, user: schemas.UserCreate):
+    if not get_user_by_slack_id_and_email(db, user.slack_id, user.email):
+        raise ValueError("User does not exist")
+
     db_user = db.query(models.User).filter(models.User.slack_id == user.slack_id, models.User.email == user.email).first()
     db.delete(db_user)
     db.commit()
     return db_user
 
 def update_user(db: Session, user: schemas.UserCreate):
+    if not get_user_by_slack_id_and_email(db, user.slack_id, user.email):
+        raise ValueError("User does not exist")
+
     db_user = db.query(models.User).filter(models.User.slack_id == user.slack_id, models.User.email == user.email).first()
     db_user.name = user.name
     db.commit()
@@ -83,6 +136,9 @@ def update_user(db: Session, user: schemas.UserCreate):
     return db_user
 
 def update_token(db: Session, token: schemas.TokenCreate):
+    if not check_token(db, token.token):
+        raise ValueError("Token does not exist")
+
     db_token = db.query(models.Token).filter(models.Token.token == token.token, models.Token.owner_slack_id == token.owner_slack_id).first()
     db_token.uses_left = token.uses_left
     db.commit()
@@ -103,3 +159,11 @@ def get_tokens_by_owner_and_token_and_is_revoked(db: Session, owner_slack_id: st
 
 def get_tokens_by_owner_and_token_and_is_blocked(db: Session, owner_slack_id: str, token: str, is_blocked: bool):
     return db.query(models.Token).filter(models.Token.owner_slack_id == owner_slack_id, models.Token.token == token, models.Token.is_blocked == is_blocked).first()
+
+
+def register_usage(db: Session, usage: schemas.UsageCreate):
+    db_usage = models.Usage(**usage.dict())
+    db.add(db_usage)
+    db.commit()
+    db.refresh(db_usage)
+    return db_usage
